implemented all the stuckc

+++ data/src/main/java/me/simonpojok/data/weather/model/CoordinateDataModel.kt
@@ -1,6 +1,6 @@
 package me.simonpojok.data.weather.model
 
 data class CoordinateDataModel(
-    val lon: Int,
-    val lat: Int
+    val lon: Double,
+    val lat: Double
 )
