+++ domain/src/main/java/me/simonpojok/domain/weather/model/CoordinateDomainModel.kt
@@ -1,6 +1,6 @@
 package me.simonpojok.domain.weather.model
 
 data class CoordinateDomainModel(
-    val lon: Int,
-    val lat: Int
+    val lon: Double,
+    val lat: Double
 )
