+++ remote/src/main/java/me/simonpojok/remote/service/OpenWeatherService.kt
@@ -8,8 +8,8 @@ import retrofit2.http.Path
 interface OpenWeatherService {
     @GET("/data/2.5/weather?lat={lat}&lon={lon}&appid={apiKey}")
     suspend fun getCurrentWeatherData(
-        @Path("lat") lat: Int,
-        @Path("lon") log: Int,
+        @Path("lat") lat: Double,
+        @Path("lon") log: Double,
         @Path("apiKey") apiKey: String
     ): AreaWeatherConditionRemoteModel
 }
