+++ remote/src/main/java/me/simonpojok/remote/weather/model/CoordinateRemoteModel.kt
@@ -1,6 +1,6 @@
 package me.simonpojok.remote.weather.model
 
 data class CoordinateRemoteModel(
-    val lon: Int,
-    val lat: Int
+    val lon: Double,
+    val lat: Double
 )
